indexing: extract reader/text/output modules from index.py
index.py had grown too large. Split it along its existing section
boundaries:

- reader.py: the graph read layer (GraphTerm, GraphReader + rdflib/HDT
  readers, graph_reader, load_graph)
- text.py: pure text helpers (humanize, label predicates, digests, ...)
- output.py: OutputRecord and the write_* functions

index.py keeps the textification (Textifier + materialize_records).
Behavior is unchanged -- textify output is byte-identical -- and the
four importers are repointed at the new modules.
