+++ src/frink_embeddings_web/indexing/index.py
@@ -1,259 +1,27 @@
-import hashlib
-import json
 import re
 from collections import defaultdict
-from dataclasses import asdict, dataclass
-from pathlib import Path
-from typing import Any, Generator, Iterable
-from typing import Literal as TypingLiteral
+from typing import Any, Generator
 
 from loguru import logger
-from rdflib import BNode, Graph, Literal, Node, URIRef
-from rdflib.namespace import RDF
-from rdflib.util import from_n3
-from rdflib_hdt import HDTStore
+from rdflib import Graph
 
 from .models import (
     GraphConfiguration,
     LabelProfileConfiguration,
     MaterializationConfiguration,
 )
-
-RDFS_LABEL = "http://www.w3.org/2000/01/rdf-schema#label"
-
-
-@dataclass
-class OutputRecord:
-    iris: list[str]
-    label: str
-    embedding_text: str
-    # Total distinct IRIs that produced this text, before `iris` is capped at
-    # `max_iris_per_record`. May exceed `len(iris)`.
-    iri_count: int
-
-
-def load_graph(hdt_file: Path):
-    store = HDTStore(str(hdt_file))
-    return Graph(store=store)
-
-
-# --- graph readers --------------------------------------------------------
-#
-# A GraphReader reads triples out of a graph as lightweight `GraphTerm`s. The
-# HDT reader reads the native HDT document directly, which avoids rdflib's
-# per-term object construction -- the dominant cost when materializing. This
-# layer only reads; turning nodes into text is the Textifier's job.
-
-
-@dataclass(frozen=True)
-class GraphTerm:
-    """An RDF term as a `kind` plus its string value.
-
-    Used instead of full rdflib `URIRef`/`Literal` objects on the materialize
-    hot path: constructing/validating rdflib terms for every triple is a large
-    share of the cost, and we only ever need the string value here.
-    """
-
-    kind: TypingLiteral["iri", "literal", "bnode"]
-    value: str
-
-    def __str__(self) -> str:
-        return self.value
-
-
-class GraphReader:
-    """Read access to a graph in terms of `GraphTerm`s."""
-
-    def root_iris(self, root_type: str) -> Iterable[str]:
-        raise NotImplementedError
-
-    def predicate_objects(
-        self, subject: Any
-    ) -> Iterable[tuple[GraphTerm, GraphTerm]]:
-        raise NotImplementedError
-
-    def objects(self, subject: Any, predicate_iri: str) -> Iterable[GraphTerm]:
-        raise NotImplementedError
-
-    def term(self, node: Any) -> GraphTerm:
-        raise NotImplementedError
-
-    def types(self, node: Any) -> Iterable[str]:
-        for obj in self.objects(node, str(RDF.type)):
-            if obj.kind == "iri":
-                yield obj.value
-
-    def first_literal(
-        self, subject: Any, predicates: Iterable[str]
-    ) -> str | None:
-        for predicate in predicates:
-            for obj in self.objects(subject, predicate):
-                if obj.kind == "literal":
-                    return obj.value
-        return None
-
-
-class RDFLibGraphReader(GraphReader):
-    """Reader over any in-memory rdflib graph (used by tests)."""
-
-    def __init__(self, graph: Graph):
-        self.graph = graph
-
-    def root_iris(self, root_type: str) -> Iterable[str]:
-        for node in self.graph.subjects(RDF.type, URIRef(root_type)):
-            if isinstance(node, URIRef):
-                yield str(node)
-
-    def predicate_objects(
-        self, subject: Any
-    ) -> Iterable[tuple[GraphTerm, GraphTerm]]:
-        node = self._to_node(subject)
-        if isinstance(node, Literal):
-            return
-        for pred, obj in self.graph.predicate_objects(node):
-            if isinstance(pred, URIRef):
-                yield self.term(pred), self.term(obj)
-
-    def objects(self, subject: Any, predicate_iri: str) -> Iterable[GraphTerm]:
-        node = self._to_node(subject)
-        if isinstance(node, Literal):
-            return
-        for obj in self.graph.objects(node, URIRef(predicate_iri)):
-            yield self.term(obj)
-
-    def term(self, node: Any) -> GraphTerm:
-        if isinstance(node, GraphTerm):
-            return node
-        if isinstance(node, Literal):
-            return GraphTerm("literal", str(node))
-        if isinstance(node, BNode):
-            return GraphTerm("bnode", str(node))
-        return GraphTerm("iri", str(node))
-
-    def _to_node(self, node: Any) -> Node:
-        if isinstance(node, GraphTerm):
-            if node.kind == "literal":
-                return Literal(node.value)
-            if node.kind == "bnode":
-                return BNode(node.value)
-            return URIRef(node.value)
-        return node
-
-
-class HDTGraphReader(GraphReader):
-    """Reader that reads the native HDT document, skipping rdflib terms."""
-
-    def __init__(self, graph: Graph):
-        self.graph = graph
-        self.document = graph.store.hdt_document
-
-    def root_iris(self, root_type: str) -> Iterable[str]:
-        triples, _ = self.document.search_triples("", str(RDF.type), root_type)
-        for subject, _, _ in triples:
-            if not subject.startswith("_:"):
-                yield subject
-
-    def predicate_objects(
-        self, subject: Any
-    ) -> Iterable[tuple[GraphTerm, GraphTerm]]:
-        subject_term = self.term(subject)
-        if subject_term.kind == "literal":
-            return
-        triples, _ = self.document.search_triples(subject_term.value, "", "")
-        for _, pred, obj in triples:
-            yield GraphTerm("iri", pred), self._term_from_hdt(obj)
-
-    def objects(self, subject: Any, predicate_iri: str) -> Iterable[GraphTerm]:
-        subject_term = self.term(subject)
-        if subject_term.kind == "literal":
-            return
-        triples, _ = self.document.search_triples(
-            subject_term.value, predicate_iri, ""
-        )
-        for _, _, obj in triples:
-            yield self._term_from_hdt(obj)
-
-    def term(self, node: Any) -> GraphTerm:
-        if isinstance(node, GraphTerm):
-            return node
-        if isinstance(node, Literal):
-            return GraphTerm("literal", str(node))
-        if isinstance(node, BNode):
-            return GraphTerm("bnode", f"_:{node}")
-        if isinstance(node, URIRef):
-            return GraphTerm("iri", str(node))
-        return self._term_from_hdt(str(node))
-
-    def _term_from_hdt(self, text: str) -> GraphTerm:
-        if text.startswith("_:"):
-            return GraphTerm("bnode", text)
-        if text.startswith('"'):
-            try:
-                node = from_n3(text)
-            except Exception:
-                return GraphTerm("literal", text)
-            if isinstance(node, Literal):
-                return GraphTerm("literal", str(node))
-            return GraphTerm("literal", text)
-        return GraphTerm("iri", text)
-
-
-def graph_reader(graph: Graph) -> GraphReader:
-    if hasattr(graph.store, "hdt_document"):
-        return HDTGraphReader(graph)
-    return RDFLibGraphReader(graph)
-
-
-# --- text helpers (pure) --------------------------------------------------
-
-
-def humanize(text: str) -> str:
-    text = text.replace("_", " ").replace("-", " ")
-    out = []
-    prev_lower = False
-
-    for c in text:
-        if prev_lower and c.isupper():
-            out.append(" ")
-        out.append(c)
-        prev_lower = c.islower()
-
-    return " ".join("".join(out).split())
-
-
-def iri_fragment(iri: str) -> str:
-    if "#" in iri:
-        return iri.rsplit("#", 1)[1]
-    if "/" in iri:
-        return iri.rsplit("/", 1)[1]
-    return iri
-
-
-def fallback_label(iri: str) -> str:
-    return humanize(iri_fragment(iri))
-
-
-def effective_label_predicates(config: GraphConfiguration) -> list[str]:
-    predicates = [*config.label_predicates]
-    if config.include_rdfs_label and RDFS_LABEL not in predicates:
-        predicates.append(RDFS_LABEL)
-    return predicates
-
-
-def normalize_label(text: str) -> str:
-    return " ".join(text.split())
-
-
-def stable_score(root: Any, pred: Any, obj: Any) -> str:
-    text = f"{root}\t{pred}\t{obj}"
-    return hashlib.sha256(text.encode("utf-8")).hexdigest()
-
-
-def text_digest(text: str) -> str:
-    return hashlib.sha256(text.encode("utf-8")).hexdigest()
-
-
-# --- textification --------------------------------------------------------
+from .output import OutputRecord
+from .reader import GraphReader, GraphTerm, graph_reader
+from .text import (
+    RDFS_LABEL,
+    effective_label_predicates,
+    fallback_label,
+    humanize,
+    iri_fragment,
+    normalize_label,
+    stable_score,
+    text_digest,
+)
 
 
 class Textifier:
@@ -481,7 +249,7 @@ class Textifier:
 
 
 def materialize_records(
-    graph: Graph,
+    graph: "Graph",
     config: MaterializationConfiguration,
     target: str | None = None,
     limit: int | None = None,
@@ -534,39 +302,3 @@ def materialize_records(
                 record.embedding_text,
             )
     return sorted(records, key=lambda r: (r.embedding_text, r.iris))
-
-
-# --- output writers -------------------------------------------------------
-
-
-def write_json(records: Iterable[OutputRecord], output_path: Path):
-    with output_path.open("w", encoding="utf-8") as f:
-        json.dump(
-            [asdict(r) for r in records],
-            f,
-            indent=2,
-            ensure_ascii=False,
-        )
-
-
-def write_jsonl(records: Iterable[OutputRecord], output_path: Path):
-    with output_path.open("w", encoding="utf-8") as f:
-        for record in records:
-            write_jsonl_record(record, f)
-
-
-def write_jsonl_record(record: OutputRecord, f) -> None:
-    json.dump(asdict(record), f, ensure_ascii=False)
-    f.write("\n")
-
-
-def write_text(records: Iterable[OutputRecord], output_path: Path):
-    with output_path.open("w", encoding="utf-8") as f:
-        for r in records:
-            f.write(f"label: {r.label}\n")
-            f.write("iris:\n")
-            for iri in r.iris:
-                f.write(f"- {iri}\n")
-            f.write("\n")
-            f.write(r.embedding_text)
-            f.write("\n\n---\n\n")
